requirements.txt
python-dotenv
together
streamlit==1.29.0
joblib==1.3.2
torch==2.3.1
xformers==0.0.27
peft 
accelerate 
bitsandbytes 
triton
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git